+++ pydcop/commands/generate.py
@@ -50,13 +50,15 @@ tt currently generate problems for the following types:
 
    generate/graphcoloring
    generate/meetingscheduling
+   generate/ising
 
 
 Planned
 -------
 
 * mixed constraints: generate dcops with both hard and soft constraints
-* ising problem
+* IoT-like problems
+* sensor network
 
 
 
@@ -89,7 +91,7 @@ import networkx as nx
 import os
 
 #from numpy.random import random
-from pydcop.commands.generators import graphcoloring, meetingscheduling
+from pydcop.commands.generators import graphcoloring, meetingscheduling, ising
 from pydcop.commands.generators.iot import generate_iot
 from pydcop.commands.generators.secp import generate_secp, parser_secp
 from pydcop.commands.generators.smallworld import generate_small_world
@@ -126,9 +128,9 @@ def set_parser(main_subparsers):
     # parser.set_defaults(func=run_cmd)
 
     graphcoloring.init_cli_parser(subparsers)
-    # parser_graph_coloring(subparsers)
-
     meetingscheduling.init_cli_parser(subparsers)
+    ising.init_cli_parser(subparsers)
+
 
     parser_mixed_problem(subparsers)